requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
torchaudio
numpy
pandas
opencv-python-headless
albumentations
timm
segmentation-models-pytorch
reportlab
transformers
pillow
scikit-learn
huggingface_hub
sqlalchemy
